requirements.txt
coverage==7.11.3
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.0
pytest-cov==7.0.0

# ---- CLI Cabinet ----
requests==2.32.3
pydantic-settings==2.4.0
pydantic==2.8.2
httpx>=0.27
# ---- Tests ----
pytest-cov
fastapi
uvicorn[standard]
pydantic[email]

# ---- Kafka ----
kafka-python
